refactor: added shared types and functions
